remove boards cuz too much complication

+++ app/routes/api.delete-zone.ts
@@ -3,6 +3,7 @@ import { ActionFunctionArgs } from "@remix-run/node";
 import { eq } from "drizzle-orm";
 import { db } from "server/db";
 import { pins, presetZones, zones } from "server/schema";
+import { updateZoneColor } from "server/utils";
 
 const schema = z.object({
   id: z.string(),
@@ -10,6 +11,18 @@ const schema = z.object({
 
 export async function action({ request }: ActionFunctionArgs) {
   const data = schema.parse(await request.json());
+
+  const zoneData = await db.query.zones.findFirst({
+    where: (table, { eq }) => eq(table.id, data.id),
+    with: { pins: true },
+  });
+
+  if (!zoneData) return new Response(undefined, { status: 400 });
+
+  updateZoneColor({
+    ...zoneData,
+    color: "#000000",
+  });
 
   await db.batch([
     db.delete(pins).where(eq(pins.zoneId, data.id)),
